Fix initial dump provisioning for ansible 2.1.1
Since ansible 2.1.1, results for the `stat` command are registered even
when the command is skipped. That broke our previous scheme were we
saved ourselves a couple of `stat` actions by including them in our
conditional blocks.

Our registered `latest_timestamp` at the top would be overwritten with a
dictionary containing "skip" results.

We really have to perform our stat 3 times, which is what we do in this
commit.